+++ src/helpers/animateStack.ts
@@ -24,16 +24,15 @@ export default function animateStack(
 
   const arr = Array.prototype.slice.call(items) as ((HTMLElement | SVGElement) & { _hideStyle?: string })[];
   const parent = arr[0].parentElement!;
-  parent.style.overflow = "visible"; // otherwise items is hidden under target
   let resMe: (isEnd: boolean) => void;
 
   // calc hideStyle position or use previous saved: when showing>partially show> hidding
   if (arr[0]._hideStyle === undefined) {
+    parent.style.overflow = "visible"; // otherwise items is hidden under target
+    parent.style.zIndex = ((Number.parseInt(getComputedStyle(target).zIndex, 10) || 1) - 1).toString(); // setup  zIndex to be hidden under target
     const r0 = target.getBoundingClientRect();
-    const zi = ((Number.parseInt(getComputedStyle(target).zIndex, 10) || 1) - 1).toString();
     arr.forEach((a) => {
       const r = a.getBoundingClientRect();
-      a.style.zIndex = zi; // setup  zIndex to be hidden under target
       a._hideStyle = isVertical // WARN: possible issue if opening horizontal > stop > hidding vertical
         ? `translateY(${Math.round(r0.top - r.top)}px)`
         : `translateX(${Math.round(r0.left - r.left)}px)`;
@@ -45,11 +44,11 @@ export default function animateStack(
 
   const reset = (): void => {
     // reset styles
+    parent.style.overflow = "";
+    parent.style.zIndex = "";
     arr.forEach((a) => {
       delete a._hideStyle;
-      a.style.zIndex = "";
       a.style.transition = "";
-      parent.style.overflow = "";
     });
     setTimeout(() => arr.forEach((a) => (a.style.transform = "")), 1); // timeout to avoid possible blink effect when need to hide
   };
@@ -57,8 +56,6 @@ export default function animateStack(
   let tid: ReturnType<typeof setTimeout>;
   window.requestAnimationFrame(() => {
     arr.forEach((a) => {
-      // todo need extra option to animate step by step so 3 goes to 1st position, 2 goes to 2nd etc...
-      // console.warn("=>", a.outerHTML);
       a.style.transition = `transform ${ms}ms ease-out`;
       a.style.transform = isHide ? a._hideStyle! : "";
     });